Moved workflows to separate package.
