+++ src/citrine/informatics/workflows/__init__.py
@@ -1,0 +1,3 @@
+from .design_workflow import *
+from .performance_workflow import *
+from .predictor_evaluation_workflow import *
